Creates ObacoService network service class
This class will handle most calls to the alerts.onebusaway.org/onebusaway.co service which do not depend on a specific region being available.
