+++ OBAKit/Services/PromisedModelService.swift
@@ -29,99 +29,18 @@ import PromiseKit
         let escapedStopID = OBAURLHelpers.escapePathVariable(stopID)
         let path = String.init(format: "/api/where/arrivals-and-departures-for-stop/%@.json", escapedStopID)
 
-        return self.obaJsonDataSource.buildGETRequest(withPath: path, queryParameters: args)
+        return obaJsonDataSource.buildGETRequest(withPath: path, queryParameters: args)
     }
 
     private func decodeStopArrivals(json: Any) throws -> OBAArrivalsAndDeparturesForStopV2 {
         var error: NSError?
 
-        let modelObjects = self.modelFactory.getArrivalsAndDeparturesForStopV2(fromJSON: json as! [AnyHashable: Any], error: &error)
+        let modelObjects = modelFactory.getArrivalsAndDeparturesForStopV2(fromJSON: json as! [AnyHashable: Any], error: &error)
         if let error = error {
             throw error
         }
 
         return modelObjects
-    }
-}
-
-// MARK: - Weather
-@objc extension PromisedModelService {
-
-    /// Request the forecasted weather for the user's region and/or location.
-    ///
-    /// - Parameters:
-    ///   - region: The user's current region
-    ///   - location: An optional location used to determine more accurate weather data.
-    /// - Returns: A promise wrapper that resolves to a WeatherForecast object.
-    @objc public func requestWeather(in region: OBARegionV2, location: CLLocation?) -> PromiseWrapper {
-        let request = buildURLRequestForWeather(in: region, location: location)
-        let wrapper = PromiseWrapper(request: request, dataDecodingStrategy: .noStrategy)
-
-        wrapper.promise = wrapper.promise.then { networkResponse -> NetworkResponse in
-            let forecast = try self.decodeWeather(data: networkResponse.object as! Data)
-            return NetworkResponse.init(object: forecast, URLResponse: networkResponse.URLResponse, urlRequest: networkResponse.urlRequest)
-        }
-
-        return wrapper
-    }
-
-    @nonobjc private func decodeWeather(data: Data) throws -> WeatherForecast {
-        let forecast = try WeatherForecast.decoder.decode(WeatherForecast.self, from: data)
-        return forecast
-    }
-
-    @nonobjc private func buildURLRequestForWeather(in region: OBARegionV2, location: CLLocation?) -> OBAURLRequest {
-        let path = "/api/v1/regions/\(region.identifier)/weather"
-
-        var params: [AnyHashable: Any] = [:]
-        if let location = location {
-            params["lat"] = location.coordinate.latitude
-            params["lng"] = location.coordinate.longitude
-        }
-
-        return obacoJsonDataSource.buildGETRequest(withPath: path, queryParameters: params)
-    }
-}
-
-// MARK: - Alarms
-@objc extension PromisedModelService {
-    /// Creates an alarm object on the server
-    ///
-    /// - Parameters:
-    ///   - alarm: The local alarm object
-    ///   - userPushNotificationID: The user's unique push notification ID
-    /// - Returns: A promise that fulfills into an URL object
-    @objc func createAlarm(_ alarm: OBAAlarm, userPushNotificationID: String) -> PromiseWrapper {
-        let request = createAlarmRequest(alarm, userPushNotificationID: userPushNotificationID)
-        let wrapper = PromiseWrapper.init(request: request)
-
-        wrapper.promise = wrapper.promise.then { networkResponse -> NetworkResponse in
-            let dict = networkResponse.object as! [String: Any]
-
-            if dict["error"] != nil {
-                throw OBAErrorMessages.cannotRegisterAlarm
-            }
-
-            let url = URL.init(string: dict["url"] as! String)!
-
-            return NetworkResponse.init(object: url, URLResponse: networkResponse.URLResponse, urlRequest: networkResponse.urlRequest)
-        }
-
-        return wrapper
-    }
-
-    @nonobjc private func createAlarmRequest(_ alarm: OBAAlarm, userPushNotificationID: String) -> OBAURLRequest {
-        let params: [String: Any] = [
-            "seconds_before": alarm.timeIntervalBeforeDeparture,
-            "stop_id": alarm.stopID,
-            "trip_id": alarm.tripID,
-            "service_date": alarm.serviceDate,
-            "vehicle_id": alarm.vehicleID,
-            "stop_sequence": alarm.stopSequence,
-            "user_push_id": userPushNotificationID
-        ]
-
-        return self.obacoJsonDataSource.buildRequest(withPath: "/api/v1/regions/\(alarm.regionIdentifier)/alarms", httpMethod: "POST", queryParameters: nil, formBody: params)
     }
 }
 
@@ -145,7 +64,7 @@ import PromiseKit
 
     @nonobjc private func decodeTripDetails(json: [AnyHashable: Any]) throws -> OBATripDetailsV2 {
         var error: NSError?
-        let model = self.modelFactory.getTripDetailsV2(fromJSON: json, error: &error)
+        let model = modelFactory.getTripDetailsV2(fromJSON: json, error: &error)
 
         if let error = error {
             throw error
@@ -167,7 +86,7 @@ import PromiseKit
 
         let escapedTripID = OBAURLHelpers.escapePathVariable(tripInstance.tripId)
 
-        return self.obaJsonDataSource.buildGETRequest(withPath: "/api/where/trip-details/\(escapedTripID).json", queryParameters: args)
+        return obaJsonDataSource.buildGETRequest(withPath: "/api/where/trip-details/\(escapedTripID).json", queryParameters: args)
     }
 }
 
@@ -227,7 +146,7 @@ extension PromisedModelService {
                     }.compactMap { try? AgencyAlert(feedEntity: $0, agencies: agencies) }
                     acc.append(contentsOf: alerts)
                 })
-                return Promise.init(value: allAlerts)
+                return Promise(value: allAlerts)
             }
         }
     }
@@ -246,66 +165,5 @@ extension PromisedModelService {
         let encodedID = OBAURLHelpers.escapePathVariable(agency.agencyId)
         let path = "/api/gtfs_realtime/alerts-for-agency/\(encodedID).pb"
         return unparsedDataSource.buildGETRequest(withPath: path, queryParameters: nil)
-    }
-}
-
-// MARK: - Vehicle Search
-@objc extension PromisedModelService {
-    /// Returns a PromiseWrapper that resolves to an array of `MatchingAgencyVehicle` objects,
-    /// suitable for passing along to `requestVehicleTrip()`.
-    ///
-    /// - Parameter matching: A substring that must appear in all returned vehicles
-    /// - Parameter region: The region from which to load all vehicle IDs
-    /// - Returns: A `PromiseWrapper` that resolves to `[MatchingAgencyVehicle]`
-    @objc public func requestVehicles(matching: String, in region: OBARegionV2) -> PromiseWrapper {
-        let request = buildVehicleListRequest(matching: matching, region: region)
-        let wrapper = PromiseWrapper(request: request, dataDecodingStrategy: .noStrategy)
-
-        wrapper.promise = wrapper.promise.then { networkResponse -> NetworkResponse in
-            let decoder = JSONDecoder()
-            let objects = try decoder.decode([MatchingAgencyVehicle].self, from: networkResponse.object as! Data)
-
-            if objects.count == 0 {
-                throw OBAErrorMessages.vehicleNotFoundError
-            }
-
-            return NetworkResponse.init(object: objects, URLResponse: networkResponse.URLResponse, urlRequest: networkResponse.urlRequest)
-        }
-
-        return wrapper
-    }
-
-    private func buildVehicleListRequest(matching: String, region: OBARegionV2) -> OBAURLRequest {
-        let path = "/api/v1/regions/\(region.identifier)/vehicles"
-        let url = obacoJsonDataSource.constructURL(fromPath: path, params: ["query": matching])
-        let obacoRequest = OBAURLRequest.init(url: url, cachePolicy: .useProtocolCachePolicy, timeoutInterval: 10)
-        return obacoRequest
-    }
-
-    /// Returns a PromiseWrapper that resolves to an OBATripDetailsV2 object.
-    ///
-    /// - Parameter vehicleID: The vehicle for which to retrieve trip details.
-    /// - Returns: a PromiseWrapper that resolves to trip details.
-    @objc public func requestVehicleTrip(_ vehicleID: String) -> PromiseWrapper {
-        let request = buildTripForVehicleRequest(vehicleID)
-        let wrapper = PromiseWrapper(request: request)
-
-        wrapper.promise = wrapper.promise.then { response -> NetworkResponse in
-            var error: NSError?
-            let entryWithRefs = self.modelFactory.getTripDetailsV2(fromJSON: response.object as! [AnyHashable: Any], error: &error)
-            if let error = error { throw error }
-
-            let tripDetails = entryWithRefs.entry as! OBATripDetailsV2
-
-            return NetworkResponse.init(object: tripDetails, response: response)
-        }
-
-        return wrapper
-    }
-
-    private func buildTripForVehicleRequest(_ vehicleID: String) -> OBAURLRequest {
-        let encodedID = OBAURLHelpers.escapePathVariable(vehicleID)
-        let path = "/api/where/trip-for-vehicle/\(encodedID).json"
-        return obaJsonDataSource.buildGETRequest(withPath: path, queryParameters: nil)
     }
 }
