+++ OneBusAway/ui/map/OBAMapViewController.m
@@ -273,12 +273,12 @@ static const double kStopsInRegionRefreshDelayOnDrag = 0.1;
     else if ([target isKindOfClass:OBAVehicleIDNavigationTarget.class]) {
         [SVProgressHUD show];
         OBAVehicleIDNavigationTarget *vehicleNavTarget = (OBAVehicleIDNavigationTarget*)target;
-        PromiseWrapper *wrapper = [self.modelService requestVehiclesMatching:vehicleNavTarget.query in:self.modelDAO.currentRegion];
+        PromiseWrapper *wrapper = [self.application.obacoService requestVehiclesMatching:vehicleNavTarget.query in:self.modelDAO.currentRegion];
         wrapper.anyPromise.then(^(NetworkResponse *response) {
             NSArray<OBAMatchingAgencyVehicle*> *matchingVehicles = response.object;
 
             if (matchingVehicles.count == 1) {
-                return [self.modelService requestVehicleTrip:matchingVehicles.firstObject.vehicleID].anyPromise;
+                return [self.application.obacoService requestVehicleTrip:matchingVehicles.firstObject.vehicleID].anyPromise;
             }
             else {
                 // pop up a disambiguation UI.
@@ -1035,9 +1035,7 @@ static const double kStopsInRegionRefreshDelayOnDrag = 0.1;
     [viewController dismissViewControllerAnimated:YES completion:^{
         [SVProgressHUD show];
 
-        // abxoxo - todo - add this wrapper to a 'disposal bag' or something that can be cancelled
-        // if the user exits this view controller before this operation finishes.
-        PromiseWrapper *wrapper = [self.modelService requestVehicleTrip:matchingVehicle.vehicleID];
+        PromiseWrapper *wrapper = [self.application.obacoService requestVehicleTrip:matchingVehicle.vehicleID];
         wrapper.anyPromise.then(^(NetworkResponse *response){
             OBATripDetailsV2 *tripDetails = (OBATripDetailsV2 *)response.object;
             OBAArrivalAndDepartureViewController *controller = [[OBAArrivalAndDepartureViewController alloc] initWithTripInstance:tripDetails.tripInstance];
